Added support for VS Code Remote-Containers extension
When running with a remote development environment the extension defaults to running in the VS Code Server on the remote container/server, this prevents it from connecting with the local browser. This change forces VS code to run in the UI environment rather than the server environment inside the container.